requirements.txt
absl-py==0.9.0
apparmor==2.13.3
appdirs==1.4.3
apsw==3.30.1.post1
asn1crypto==0.24.0
astropy==4.0
astunparse==1.6.3
atomicwrites==1.1.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.4
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
configobj==5.0.6
CredentialKit==0.7
credentialkit-client==1
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.3.0
defusedxml==0.6.0
distro==1.4.0
distro-info==0.23
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.11.0
entrypoints==0.3
evdev==1.3.0
ExifRead==2.1.2
extras==1.0.0
fasteners==0.14.1
feedparser==5.2.1
fixtures==3.0.0
future==0.18.2
gast==0.3.3
glinux-homedir-helper==1
glinux-identity==1
glinux-welcome==7
goobuntu-config-tools==0.1
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googlenetworkaccess==0.1
grpcio==1.29.0
guizero==1.1.0
h5py==2.10.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
hypothesis==4.36.2
idna==2.8
ifaddr==0.1.6
importlib-metadata==1.5.0
iniparse==0.4
IPy==0.83
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.16.0
jeepney==0.4.3
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==6.1.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.1.2
kernel-pruner==52
keyring==18.0.1
keyrings.alt==3.2.0
keyutils==0.6
kiwisolver==1.0.1
LibAppArmor==2.13.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.13.0
lxml==4.5.0
Mako==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mechanize==0.4.5
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mox3==0.28.0
mpi4py==3.0.3
msgpack==0.5.6
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.9
nose==1.3.7
notebook==6.0.3
nudatus==0.0.4
numpy==1.17.4
numpy-stl==2.11.2
oauthlib==3.1.0
obno==39
olefile==0.46
onboard==1.4.1
opt-einsum==3.2.1
packaging==20.1
pandocfilters==1.4.2
parameterized==0.7.0
paramiko==2.6.0
parso==0.6.2
pbr==5.4.3
pexpect==4.6.0
pgzero==1.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.12.1
psutil==5.6.7
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pychm==0.8.6
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.4.1
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.3.1
PyGObject==3.34.0
pygpu==0.7.6
pyinotify==0.9.6
PyJWT==1.7.1
PyKCS11==1.5.7
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyrsistent==0.15.5
pyserial==3.4
pysmbc==1.0.15.6
pytest==4.6.9
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-apt==1.8.6
python-augeas==0.5.0
python-dateutil==2.7.3
python-debian==0.1.36
python-mimeparse==1.6.0
python-networkmanager==2.1
python-pam==1.8.4
python-utils==2.4.0
python-xapp==1.8.1
python-xlib==0.26
pyudev==0.21.0
pyusb==1.0.2
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.3.1
reconfigure==0.1.81
regex==2019.8.19
rekey==1
repoze.lru==0.7
requests==2.22.0
requests-oauthlib==1.3.0
Routes==2.4.1
rsa==4.0
scipy==1.4.1
SecretStorage==3.1.2
semver==2.0.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
testtools==2.3.0
Theano==1.0.4
tinycss==0.4
tornado==5.1.1
traceback2==1.4.0
traitlets==4.3.3
udiskie==2.1.0
uflash==1.2.4
ufw==0.36
unittest2==1.1.0
urllib3==1.25.8
virtualenv==15.1.0
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==1.0.1
widgetsnbextension==2.0.0
wrapt==1.12.1
youtube-dl==2020.1.24
zeroconf==0.23.0
zipp==1.0.0
